refactor(frontend): functions to enhance redirect
