+++ apps/frontend/app/lib/generals.ts
@@ -205,6 +205,7 @@ export const AUTH_COOKIE_NAME = "Auth";
 export const USER_PREFERENCES_COOKIE_NAME = "UserPreferences";
 export const CORE_DETAILS_COOKIE_NAME = "CoreDetails";
 export const USER_DETAILS_COOKIE_NAME = "UserDetails";
+export const SEARCH_PARAM_COOKIE = "SearchParams=";
 
 const baseUrl = typeof window !== "undefined" ? window.location.origin : "";
